fix(dropper-ui): recover media the SSE announcement never delivered
The announcement bus is lossy but the pipeline result is not: it always names the
artifact path. The dropper only rendered media it had heard announced, so a dropped
event silently cost the user their video. It now reconciles the result against what
SSE delivered and pulls the difference.

Also folds nine test files into two: the three node suites describe one behaviour
from three sides, and the spool and the command that serves it share a contract.
