+++ apps/dropper-ui/src/utils/dropperUtils.ts
@@ -98,19 +98,53 @@ const processTableData = (data: any): string[] => {
 };
 
 
-/**
- * Extract playable sources from a media lane (image/audio/video).
- *
- * The response node emits `{mime_type, path}` when a FileStore persisted the
- * bytes — those arrive live over SSE and are merged in the hook, so they are
- * skipped here. Without a store it falls back to `{mime_type, <kind>: base64}`,
- * which never announces; turn those into a data URL so they still render.
- */
+/** A persisted media artifact named in the result: `{mime_type, path}`. */
+export interface MediaArtifact {
+	filename: string;
+	kind: string;
+	name: string;
+	path: string;
+	mime: string;
+}
+
+/** Data URLs for a lane's base64 entries, emitted when no FileStore persisted the bytes. */
 const processMediaData = (fieldData: any, kind: string): string[] => {
 	if (!Array.isArray(fieldData)) return [];
 	return fieldData
 		.filter(entry => entry && typeof entry === 'object' && typeof entry[kind] === 'string')
 		.map(entry => `data:${entry.mime_type || 'application/octet-stream'};base64,${entry[kind]}`);
+};
+
+const MEDIA_LANES = new Set(['image', 'audio', 'video']);
+
+/**
+ * Every persisted media artifact named in the results.
+ *
+ * The response node announces each one over SSE, but that bus is lossy — a dropped
+ * announcement would silently cost the user their media. The result always carries
+ * the path, so the hook merges whatever SSE did not deliver.
+ */
+export const extractMediaArtifacts = (uploadResults: UPLOAD_RESULT[]): MediaArtifact[] => {
+	const artifacts: MediaArtifact[] = [];
+	for (const uploadResult of uploadResults) {
+		const result: any = uploadResult.result;
+		if (!result?.result_types) continue;
+		for (const [fieldName, fieldType] of Object.entries(result.result_types)) {
+			const kind = String(fieldType);
+			if (!MEDIA_LANES.has(kind) || !Array.isArray(result[fieldName])) continue;
+			for (const entry of result[fieldName]) {
+				if (!entry || typeof entry.path !== 'string' || !entry.path) continue;
+				artifacts.push({
+					filename: uploadResult.filepath,
+					kind,
+					name: entry.path.split('/').pop() || kind,
+					path: entry.path,
+					mime: typeof entry.mime_type === 'string' ? entry.mime_type : ''
+				});
+			}
+		}
+	}
+	return artifacts;
 };
 
 // ============================================================================
