--- packages/ai/tests/ai/account/__init__.py
@@ -1,1 +1,0 @@
-# Test package for ai.account module
