+++ packages/ai/tests/ai/test_live_media.py
@@ -1,29 +1,194 @@
-"""
-Unit tests for ai.modules.task.commands.cmd_media.MediaCommands.
+"""The live media channel: a spool the producer writes and the server reads along.
 
-Focus areas: the ``task.data`` gate (NOT ``task.store``), ``outputs/`` scoping, and
-the live-vs-finished read contract — a read at the end of a live artifact waits for
-the producing node instead of reporting EOF, and ``complete`` tells the two apart.
+Both halves in one file — ai.account.live_media and the rrext_media command that
+serves it — because the contract they share is the point: a read at the end of a
+live artifact waits for the producer, so empty bytes only ever mean end-of-stream.
 """
-
-from __future__ import annotations
 
 import asyncio
+import os
+import time
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from ai.account.live_media import LiveWriter
+from ai.account import live_media
+from ai.account.live_media import LiveReader, LiveWriter
 from ai.modules.task.commands.cmd_media import MediaCommands, _ARTIFACT_PREFIX
 
-PATH = 'outputs/video/a.mp4'
+CLIENT = 'user-1'
+PATH = 'outputs/video/x.mp4'
 
 
 @pytest.fixture(autouse=True)
 def spool_dir(tmp_path, monkeypatch):
-    """Isolate the live-media spool directory per test."""
+    """Isolate each test's spool directory."""
     monkeypatch.setenv('ROCKETRIDE_LIVE_MEDIA_DIR', str(tmp_path))
     yield tmp_path
+
+
+# ===========================================================================
+# The spool
+# ===========================================================================
+
+
+def _reader() -> LiveReader:
+    r = LiveReader(CLIENT, PATH)
+    r.open()
+    return r
+
+
+@pytest.mark.asyncio
+async def test_read_waits_for_the_producer_instead_of_reporting_eof():
+    """The core of the design: offset == available is 'not yet', not 'the end'."""
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    w.append(b'first')
+    r = _reader()
+
+    async def produce_later():
+        await asyncio.sleep(0.1)
+        w.append(b'second')
+
+    try:
+        assert await r.read(0, 99) == b'first'
+        task = asyncio.create_task(produce_later())
+        # Would be b'' (EOF) under the old semantics; must block until 'second' lands.
+        assert await r.read(5, 99, timeout=5) == b'second'
+        await task
+    finally:
+        r.close()
+        w.discard()
+
+
+@pytest.mark.asyncio
+async def test_eof_only_after_finish():
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    w.append(b'abc')
+    r = _reader()
+    try:
+        assert await r.read(0, 99) == b'abc'
+        assert not r.complete()
+        w.finish()
+        assert r.complete()
+        assert await r.read(3, 99) == b''
+    finally:
+        r.close()
+        w.discard()
+
+
+@pytest.mark.asyncio
+async def test_read_times_out_on_a_stalled_producer():
+    """A node that hangs must not hang the connection with it."""
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    w.append(b'abc')
+    r = _reader()
+    try:
+        with pytest.raises(TimeoutError, match='stalled at offset 3'):
+            await r.read(3, 99, timeout=0.1)
+    finally:
+        r.close()
+        w.discard()
+
+
+@pytest.mark.asyncio
+async def test_final_chunk_is_never_missed_when_finish_races_the_reader():
+    """finish() writes bytes then the sidecar; read() checks size then the sidecar."""
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    r = _reader()
+    try:
+        w.append(b'tail')
+        w.finish()
+        # complete() is already true — the read must still return the trailing bytes.
+        assert r.complete()
+        assert await r.read(0, 99) == b'tail'
+        assert await r.read(4, 99) == b''
+    finally:
+        r.close()
+        w.discard()
+
+
+@pytest.mark.asyncio
+async def test_rewind_while_still_producing():
+    """The stream is rewindable: a reader may seek back to bytes it already passed."""
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    w.append(b'0123456789')
+    r = _reader()
+    try:
+        assert await r.read(5, 99) == b'56789'
+        assert await r.read(0, 3) == b'012'
+    finally:
+        r.close()
+        w.discard()
+
+
+@pytest.mark.skipif(os.name == 'nt', reason='Windows cannot unlink a file held open by a reader')
+@pytest.mark.asyncio
+async def test_reader_survives_the_spool_being_reclaimed():
+    """discard() runs once the artifact is persisted; an open reader keeps working."""
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    w.append(b'payload')
+    w.finish()
+    r = _reader()
+    try:
+        w.discard()
+        assert not live_media.is_live(CLIENT, PATH)
+        assert await r.read(0, 99) == b'payload'
+    finally:
+        r.close()
+
+
+def test_discard_never_raises_when_a_reader_holds_the_spool():
+    """On Windows the unlink fails; the node must not care."""
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    w.append(b'x')
+    w.finish()
+    r = _reader()
+    try:
+        w.discard()  # must not raise on any platform
+    finally:
+        r.close()
+
+
+def test_sweep_reclaims_spools_a_crashed_node_left_behind():
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    w.append(b'orphan')  # never finished, never discarded
+
+    part, _ = live_media.spool_paths(CLIENT, PATH)
+    assert live_media.sweep_stale_spools(max_age=3600) == 0, 'a fresh spool is not stale'
+
+    old = time.time() - 7200
+    os.utime(part, (old, old))
+    assert live_media.sweep_stale_spools(max_age=3600) == 1
+    assert not live_media.is_live(CLIENT, PATH)
+
+
+def test_begin_discards_a_stale_spool():
+    w = LiveWriter(CLIENT, PATH)
+    w.begin()
+    w.append(b'old')
+    w.finish()
+
+    w2 = LiveWriter(CLIENT, PATH)
+    w2.begin()
+    try:
+        part, done = live_media.spool_paths(CLIENT, PATH)
+        assert os.path.getsize(part) == 0
+        assert not os.path.exists(done), 'a fresh stream must not inherit the old EOF marker'
+    finally:
+        w2.discard()
+
+
+# ===========================================================================
+# The rrext_media command that serves it
+# ===========================================================================
 
 
 def _make_conn(*, file_store=None, connection_id=7):
@@ -155,7 +320,7 @@ async def test_read_rejects_a_bad_offset_or_length(bad):
 @pytest.mark.asyncio
 async def test_live_artifact_opens_from_the_spool_and_is_incomplete():
     """A producing node's artifact never touches the store, and size is a snapshot."""
-    writer = LiveWriter('user-1', PATH)
+    writer = LiveWriter(CLIENT, PATH)
     writer.begin()
     writer.append(b'partial')
     fs = MagicMock()
@@ -176,7 +341,7 @@ async def test_live_artifact_opens_from_the_spool_and_is_incomplete():
 @pytest.mark.asyncio
 async def test_live_read_waits_for_the_producer_rather_than_reporting_eof():
     """The whole point: an empty chunk must never mean 'not yet'."""
-    writer = LiveWriter('user-1', PATH)
+    writer = LiveWriter(CLIENT, PATH)
     writer.begin()
     writer.append(b'first')
     conn = _make_conn()
@@ -202,7 +367,7 @@ async def test_live_read_waits_for_the_producer_rather_than_reporting_eof():
 
 @pytest.mark.asyncio
 async def test_live_read_returns_eof_only_once_the_node_finishes():
-    writer = LiveWriter('user-1', PATH)
+    writer = LiveWriter(CLIENT, PATH)
     writer.begin()
     writer.append(b'abc')
     conn = _make_conn()
@@ -221,7 +386,7 @@ async def test_live_read_returns_eof_only_once_the_node_finishes():
 
 @pytest.mark.asyncio
 async def test_live_read_is_rewindable_while_producing():
-    writer = LiveWriter('user-1', PATH)
+    writer = LiveWriter(CLIENT, PATH)
     writer.begin()
     writer.append(b'0123456789')
     conn = _make_conn()
